Refactored injection of output mode in StateStoreSaveExec
