+++ sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/IncrementalExecution.scala
@@ -17,8 +17,7 @@
 
 package org.apache.spark.sql.execution.streaming
 
-import org.apache.spark.sql._
-import org.apache.spark.sql.catalyst.planning.PhysicalAggregation
+import org.apache.spark.sql.{OutputMode, SparkSession}
 import org.apache.spark.sql.catalyst.plans.logical._
 import org.apache.spark.sql.catalyst.rules.Rule
 import org.apache.spark.sql.execution.{QueryExecution, SparkPlan, SparkPlanner, UnaryExecNode}
@@ -35,36 +34,15 @@ class IncrementalExecution private[sql](
     val currentBatchId: Long)
   extends QueryExecution(sparkSession, logicalPlan) {
 
+  // TODO: make this always part of planning.
+  val stateStrategy = sparkSession.sessionState.planner.StatefulAggregationStrategy :: Nil
+
   // Modified planner with stateful operations.
   override def planner: SparkPlanner =
     new SparkPlanner(
       sparkSession.sparkContext,
       sparkSession.sessionState.conf,
-      Nil) {
-
-      override def strategies: Seq[Strategy] = {
-        StatefulAggregationStrategy +: super.strategies
-      }
-
-      /**
-       * Used to plan aggregation queries that are computed incrementally as part of a
-       * [[org.apache.spark.sql.ContinuousQuery]].
-       */
-      object StatefulAggregationStrategy extends Strategy {
-        override def apply(plan: LogicalPlan): Seq[SparkPlan] = plan match {
-          case PhysicalAggregation(
-            namedGroupingExpressions, aggregateExpressions, rewrittenResultExpressions, child) =>
-            execution.aggregate.Utils.planStreamingAggregation(
-              namedGroupingExpressions,
-              aggregateExpressions,
-              rewrittenResultExpressions,
-              outputMode,
-              planLater(child))
-
-          case _ => Nil
-        }
-      }
-    }
+      stateStrategy)
 
   /**
    * Records the current id for a given stateful operator in the query plan as the `state`
@@ -74,17 +52,19 @@ class IncrementalExecution private[sql](
 
   /** Locates save/restore pairs surrounding aggregation. */
   val state = new Rule[SparkPlan] {
+
     override def apply(plan: SparkPlan): SparkPlan = plan transform {
-      case StateStoreSaveExec(keys, None, outputMode,
+      case StateStoreSaveExec(keys, None, None,
              UnaryExecNode(agg,
                StateStoreRestoreExec(keys2, None, child))) =>
         val stateId = OperatorStateId(checkpointLocation, operatorId, currentBatchId)
+        val returnAllStates = if (outputMode == OutputMode.Complete) true else false
         operatorId += 1
 
         StateStoreSaveExec(
           keys,
           Some(stateId),
-          outputMode,
+          Some(returnAllStates),
           agg.withNewChildren(
             StateStoreRestoreExec(
               keys,
